modelService/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
numpy
opencv-python
torch
torchvision
facenet-pytorch
pillow
